+++ lib/mainImpl.dart
@@ -22,6 +22,7 @@ import 'package:flutter/foundation.dart';
 import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/AppReview.dart';
 import 'package:illinois/service/Appointments.dart';
+import 'package:illinois/service/Assistant.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Canvas.dart';
 import 'package:illinois/service/CustomCourses.dart';
@@ -151,7 +152,7 @@ void mainImpl({ rokwire.ConfigEnvironment? configEnvironment }) async {
       Appointments(),
       MTD(),
       SpeechToText(),
-    //Assistant(),
+      Assistant(),
       MobileAccess(),
     ]);